Move CastVote utils to the CastVoteMsg module
