+++ lib/cast_vote_msg.js
@@ -3,7 +3,6 @@ const Vote = require('./entities/sys/vote')
 const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
 const { ValidationError, NotFoundError } = require('./errors')
-const CastVote = require('./cast_vote')
 const GetVoteStats = require('./get_vote_stats')
 const SavePollRating = require('./save_poll_rating')
 const Reply = require('./reply')
@@ -22,7 +21,7 @@ module.exports = function (opts = {}) {
       // TODO: Implement transactions.
       //
 
-      await CastVote.castVote({
+      await castVote({
         vote_type,
         poll_id,
         voter_id,
@@ -94,6 +93,44 @@ module.exports = function (opts = {}) {
     }, { stripUnknown: true })
 
     return validate(msg)
+  }
+
+
+  async function castVote(args, ctx, opts = {}) {
+    Assert.object(args, 'args')
+    Assert.object(ctx, 'ctx')
+    Assert.object(opts, 'opts')
+
+    const seneca = fetchProp(ctx, 'seneca')
+
+    const voter_id = fetchProp(args, 'voter_id')
+    const voter_type = fetchProp(args, 'voter_type')
+    const poll_id = fetchProp(args, 'poll_id')
+    const vote_type = fetchProp(args, 'vote_type')
+    const vote_kind = fetchProp(args, 'vote_kind')
+    const vote_code = fetchProp(args, 'vote_code')
+
+    const poll = await seneca.make('sys/poll').load$(poll_id)
+
+    if (!poll) {
+      throw new NotFoundError('poll')
+    }
+
+    const vote_attributes = {
+      poll_id,
+      voter_id,
+      voter_type,
+      type: vote_type,
+      kind: vote_kind,
+      code: vote_code,
+      created_at: new Date()
+    }
+
+    const _new_vote = seneca.make('sys/vote')
+      .data$(vote_attributes)
+      .save$()
+
+    return
   }
 }
